pyproject.toml
[project]
name = "machado-oraculo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.0",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-google-genai>=2.1.0",
    "langgraph>=1.0.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "loguru>=0.7.0", 
]
